+++ test/lib/perfect_shape/test_square.rb
@@ -310,5 +310,18 @@ describe PerfectShape do
       _(shape.contain?(point)).must_equal false
       _(shape.contain?(*point)).must_equal shape.contain?(point)
     end
+    
+    it 'returns edges of square' do
+      shape = PerfectShape::Square.new(x: 2, y: 3, length: 50)
+      
+      expected_edges = [
+        PerfectShape::Line.new(points: [[2, 3], [2 + 50, 3]]),
+        PerfectShape::Line.new(points: [[2 + 50, 3], [2 + 50, 3 + 50]]),
+        PerfectShape::Line.new(points: [[2 + 50, 3 + 50], [2, 3 + 50]]),
+        PerfectShape::Line.new(points: [[2, 3 + 50], [2, 3]]),
+      ]
+      
+      _(shape.edges).must_equal expected_edges
+    end
   end
 end