+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.3.3
 
 - `PerfectShape::Polygon#edges` returns edges of polygon as `PerfectShape::Line` objects
+- `PerfectShape::Rectangle#edges` returns edges of rectangle as `PerfectShape::Line` objects
+- `PerfectShape::Square#edges` returns edges of square as `PerfectShape::Line` objects
 - Rename `number` arg to `level` in `CubicBezierCurve#subdivisions(level=1)`, making it signify the level of subdivision recursion to perform.
 
 ## 0.3.2